moved response from constants to utils and add return type
