+++ src/constants/responses.ts
@@ -1,5 +1,3 @@
-import JSONResponse from "../utils/JsonResponse";
-
 export const UNKNOWN_INTERACTION = {
   error: "Unknown Interaction",
 };
@@ -16,11 +14,4 @@ export const STATUS_CHECK = {
   message: "Welcome to our discord Bot Server 👋",
 };
 
-export const DISCORD_RESPONSE = (reply: string) => {
-  return new JSONResponse({
-    type: 4,
-    data: {
-      content: reply,
-    },
-  });
-};
+export const COMMAND_NOT_FOUND = "Command Not Found";
