+++ src/controllers/commandNotFound.ts
@@ -1,5 +1,7 @@
-import { DISCORD_RESPONSE } from "../constants/responses";
+import { COMMAND_NOT_FOUND } from "../constants/responses";
+import { DISCORD_RESPONSE } from "../utils/discordResponse";
+import JSONResponse from "../utils/JsonResponse";
 
-export function commandNotFound() {
-  return DISCORD_RESPONSE("Command Not found");
+export function commandNotFound(): JSONResponse {
+  return DISCORD_RESPONSE(COMMAND_NOT_FOUND);
 }
