+++ src/controllers/helloCommand.ts
@@ -1,5 +1,6 @@
-import { DISCORD_RESPONSE } from "../constants/responses";
+import { DISCORD_RESPONSE } from "../utils/discordResponse";
+import JSONResponse from "../utils/JsonResponse";
 
-export function helloCommand(userId: number) {
+export function helloCommand(userId: number): JSONResponse {
   return DISCORD_RESPONSE(`Hello <@${userId}>`);
 }
